metrics-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
shapely
rasterio
pyproj
python-dotenv
httpx